django-rest/requirements.txt
Django==2.1.*
djangorestframework
django-cors-headers
python-ldap
